Add lookahead for '{' & '/' to split on special chars
Currently (as of this commit), this fails because it isn't valid to
begin a field selector with a hyphen (-) or curly brace ({) unless it's
followed by a number, so you can't split on these characters. This adds
a check to see if the initial character is followed by an opening curly
brace or slash (regex) to allow whatever the initial character is to be
treated as the separator.

An additional specialchars test is added, though it's probably not
comprehensive.